+++ lib/src/extensions.dart
@@ -18,3 +18,10 @@ extension ResponsiveNum on num {
 
   double r(BuildContext context) => ResponsiveX.scale(context, toDouble());
 }
+
+extension ResponsiveContext on BuildContext {
+  bool get isMobile => MediaQuery.of(this).size.width < 768;
+  bool get isTablet =>
+      MediaQuery.of(this).size.width >= 768 && MediaQuery.of(this).size.width < 1024;
+  bool get isDesktop => MediaQuery.of(this).size.width >= 1024;
+}